+++ CHH_PJT2/Form/frmMember.cs
@@ -251,21 +251,29 @@ namespace CHH_PJT2
         {
             try
             {
-                MemberDB db = new MemberDB();
-                SetMemberText st = new SetMemberText(txtID.Text, txtName.Text, rbtnGenderM.Tag.ToString(), txtNum.Text, txtPostalCode.Text, txtPriAddress.Text, txtDetAddress.Text, txtSchool.Text, txtGrade.Text, dtpBirth.Value, txtlessonCode.Text, dtpRegDate.Value, txtPPNum.Text, txtNs.Text, txtAntecedent.Text);
-                bool bFlag;
-                bFlag = db.DeleteMember(st);
-
-                if (bFlag)
+                frmDelete frm = new frmDelete();
+                if (frm.ShowDialog() == DialogResult.OK)
                 {
-                    MessageBox.Show("정보 삭제가 완료되었습니다.");
-                    SetData();
+                    MemberDB db = new MemberDB();
+                    SetMemberText st = new SetMemberText(txtID.Text, txtName.Text, rbtnGenderM.Tag.ToString(), txtNum.Text, txtPostalCode.Text, txtPriAddress.Text, txtDetAddress.Text, txtSchool.Text, txtGrade.Text, dtpBirth.Value, txtlessonCode.Text, dtpRegDate.Value, txtPPNum.Text, txtNs.Text, txtAntecedent.Text);
+                    bool bFlag;
+                    bFlag = db.DeleteMember(st);
+
+                    if (bFlag)
+                    {
+                        MessageBox.Show("정보 삭제가 완료되었습니다.");
+                        SetData();
+                    }
+                    else
+                    {
+                        MessageBox.Show("형식을 잘못 입력하였거나 빈 항목이 있습니다.");
+                    }
+                    db.Dispose();
                 }
                 else
                 {
-                    MessageBox.Show("형식을 잘못 입력하였거나 빈 항목이 있습니다.");
+                    frm.Close();
                 }
-                db.Dispose();
             }
             catch (Exception err)
             {
